finish endpoints

+++ app/Http/Controllers/GoodsByLetterController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\GoodsByLetter;
+use App\Models\InstructionCards;
+use App\Models\Mercancia;
 use Illuminate\Http\Request;
 
 class GoodsByLetterController extends Controller
@@ -29,7 +31,30 @@ class GoodsByLetterController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $exist = InstructionCards::find($request->input('id_instruction_card'));
+        if(!$exist)
+            return response()->json( [
+                'message' => 'invalid instruction card id',
+            ], 400 );
+
+        $exist = Mercancia::find($request->input('id_mercancia'));
+        if(!$exist)
+            return response()->json( [
+                'message' => 'invalid mercancia id',
+            ], 400 );
+
+        $goods_by_letter = new GoodsByLetter([
+            'id_instruction_card' => $request->input('id_instruction_card'),
+            'id_mercancia' => $request->input('id_mercancia'),
+            'quantity' => $request->input('quantity'),
+        ]);
+
+        $goods_by_letter->save();
+
+        return response()->json([
+            'message' => 'Created',
+            'data' => $goods_by_letter
+        ], 200 );
     }
 
     /**
@@ -40,7 +65,12 @@ class GoodsByLetterController extends Controller
      */
     public function show($id)
     {
-        $letter = GoodsByLetter::findOrFail($id);
+        $letter = GoodsByLetter::find($id);
+
+        if(!$letter)
+            return response()->json( [
+                'message' => 'not found'
+            ], 404 );
 
         return response()->json( [
             'data' => $letter
@@ -56,7 +86,35 @@ class GoodsByLetterController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $letter = GoodsByLetter::find($id);
+
+        $exist = InstructionCards::find($request->input('id_instruction_card'));
+        if(!$exist)
+            return response()->json( [
+                'message' => 'invalid instruction card id',
+            ], 400 );
+
+        $exist = Mercancia::find($request->input('id_mercancia'));
+        if(!$exist)
+            return response()->json( [
+                'message' => 'invalid mercancia id',
+            ], 400 );
+
+        if(!$letter)
+            return response()->json( [
+                'message' => 'not found'
+            ], 404 );
+
+        $letter->update([
+            'id_instruction_card' => $request->input('id_instruction_card') ?? $letter->id_instruction_card,
+            'id_mercancia' => $request->input('id_mercancia') ?? $letter->id_mercancia,
+            'quantity' => $request->input('quantity') ?? $letter->quantity,
+        ]);
+
+        return response()->json( [
+            'message' => 'Updated',
+            'data' => $letter
+        ], 200 );
     }
 
     /**
@@ -67,7 +125,13 @@ class GoodsByLetterController extends Controller
      */
     public function destroy($id)
     {
-        $letter = GoodsByLetter::findOrFail($id);
+        $letter = GoodsByLetter::find($id);
+
+        if(!$letter)
+            return response()->json( [
+                'message' => 'not found'
+            ], 404 );
+
         $letter->delete();
 
         return response()->json( [
